Updated pulse LED to turn off if no pulse
Previously the pulse LED shined white if there was no pulse and red if there was a pulse. Now, the code turns the pulse LED off if there is no pulse so that there isn't a super bright white light blinding you.

Also took out comments that weren't needed anymore.