Fixed issue of time counter negative at -32000
The time counter for the heart rate calculation was initially an int, and because the int was overflowing, it went negative rather than continuing to add up. Changed the heart rate time counter into an unsigned long (to have more possible values) and then changed it into a float because having the heart rate values as an int or long may be preventing the value from being shown because it gets cut off. 

The issue of the heart rate calculation being zero is still present though.